ResourcePackager now only packages resources. The 4 steps to load elm, build epxressions, compile and package resources moved to PackagerCliProgram
